+++ front/src/components/organisms/Navbar.tsx
@@ -172,7 +172,7 @@ export const Navbar = ({
       >
         {isSettingsOpen && <SettingsManager />}
         <FlexBox>
-          <StyledButton onClick={() => setIsSettingsOpen(false)}>
+          <StyledButton outline onClick={() => setIsSettingsOpen(false)}>
             {t('Cerrar')}
           </StyledButton>
         </FlexBox>